requirements.txt
asgiref==3.2.7
Django==3.0.5
django-js-asset==1.2.2
django-mptt==0.11.0
pytz==2019.3
sqlparse==0.3.1